lima.nix: add @wheel to nix.settings.trusted-users
By adding the wheel group to trusted-users, we enable various
remote deployment tools to deploy to a newly started instance
without a separate configuration step.

We use the wheel group because at the time the image is built
we don't know the username the Lima will configure for SSH access
to the instance.